Add some props
Props are a way to attach data relevant to a component – I am not
too clear on this yet, but I suppose it creates a binding with
the corresponding HTML element in the DOM as well so that whenever
this data is manipulated anywhere in the app, it gets updated
in the rendered view everywhere it is used.

More to come on this as I gain more understanding of this aspect.