Add formatting for price
The price is originally stored in cents (as available in mock data)
and the starter files provide a helper function to format our price
fields. We use the formatPrice function provided in helpers.js here
to format the price to nicely display a dollar sign and 2 decimal
places. See helpers.js for implementation details of formatPrice()
function.